qt-brisa: fix threads pool
The thing is that thread pool never works for brisa.
qt rule is simple: every signal receiver receives its signal
in event loop and thread, where receiver was created.

So, we should create signal receiver in httpsessionmanager::run
method.

Also, we should be very careful with passing 'this' as parent to some
QObject, because Qt does not like when parent and object were created
in different threads.

So, remove some 'this' params from ctr to avoid wrong paranting.

Also, we should move to thread httpsessionmanager to itself.
Because, httpsessionmanager creates many other objects again with 'this'
parent param. This is totally safe for httpsessionmanager, because
we manage this objects ourselves.