qt-brisa: change logic of handling main brisa device description requests
GET /device_name.xml HTTP/1.0

brisa should response with xml which describes this device.
this xml is being generated only once.
previously brisawebfile was used to return its content.
this patch changes brisawebfile on brisawebstaticcontent,
because this xml is actually a buffer, not a file.

also use setContentType(text/xml) as default type for
this xml.