qt-brisa: fix broken by commit 281b899 `connection close`
I broke closing connection on http version 1.0 and
connection-type header, which contains 'close'.

Faulty commit is 28ab899.

In this commit is reordered 'lastRequest' inition and set it
after preparing response, this is wrong, because last request
will be used directly in this hierarchy of calls.

Actually it is a bad idea to copy HttpRequest, because it internally
shares pointer on QIODevice which cannot be safely copied, so new copy
of HttpRequest takes ownership of this request and further attempts
to read body of original request will fail, because its pointer now is NULL.

So, solution is simple: do not use last request object, but simply
pass this request in all 'prepare response' calls.